requirements.txt
django==4.2.6
flake8==6.1.0
django-ckeditor==6.7.0